requirements.txt
python-dotenv==1.0.1
gradio==4.39.0
numpy==1.26.4
soundfile==0.12.1
deep-translator==1.11.4
groq==0.15.0
SpeechRecognition==3.13.0
pydub==0.25.1
assemblyai==0.31.0
langdetect==1.0.9
gTTS==2.5.4
elevenlabs==1.50.3
huggingface-hub==0.24.6
google-generativeai==1.0.9